Trim whitespace from the parsed JWT
Since the JWT is read from a file that may be produced programmatically
(in one form or another), it is possible for there to be one or more
additional whitespace surrounding the JWT in the decorated file. This
simply ensure the whitespace is trimmed as is performed when parsing the
seed value.

Signed-off-by: Byron Ruth <b@devel.io>